feat(BE-40):user can view the submission details of their applications

+++ Apply-For-Me-Api/src/main/java/com/hydraulic/applyforme/repository/ProfessionalRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ProfessionalRepository {
     List<Professional> getAll();
 
-    List<Professional> getAll(Long pageOffset);
+    List<Professional> getAll(Integer pageOffset);
 
     Professional getOne(Long id);
